+++ tests/test_client/rpc.test.cpp
@@ -44,9 +44,9 @@
 #include <doctest/doctest.h>
 
 #if defined(RPC_HPP_ENABLE_BITSERY)
-constexpr uint64_t bitsery_adapter::config::max_func_name_size = 30;
-constexpr uint64_t bitsery_adapter::config::max_string_size = 2'048;
-constexpr uint64_t bitsery_adapter::config::max_container_size = 1'000;
+constexpr size_t bitsery_adapter::config::max_func_name_size = 30;
+constexpr size_t bitsery_adapter::config::max_string_size = 2'048;
+constexpr size_t bitsery_adapter::config::max_container_size = 1'000;
 #endif
 
 template<typename Serial>
