+++ tests/test_server/rpc.server.cpp
@@ -63,9 +63,9 @@ using rpc_hpp::adapters::boost_json_adapter;
 
 using rpc_hpp::adapters::bitsery_adapter;
 
-constexpr uint64_t bitsery_adapter::config::max_func_name_size = 30;
-constexpr uint64_t bitsery_adapter::config::max_string_size = 2'048;
-constexpr uint64_t bitsery_adapter::config::max_container_size = 1'000;
+constexpr size_t bitsery_adapter::config::max_func_name_size = 30;
+constexpr size_t bitsery_adapter::config::max_string_size = 2'048;
+constexpr size_t bitsery_adapter::config::max_container_size = 1'000;
 #endif
 
 #include <algorithm>
